最愛完成

+++ unlove_api.php
@@ -3,31 +3,28 @@
 require __DIR__ . '/__db_connect.php';
 
 $result = [];
-
-$sql ='SELECT * FROM `members_favourite` WHERE `product_sid`='.$_GET['sid'];
-$stmt=$mysqli->query($sql);
-if($stmt->num_rows>=1) {
-    $sql = "DELETE FROM `members_favourite` WHERE `member_sid`=" . $_SESSION['user']['member_sid'] . " AND `product_sid`=" . $_GET['sid'];
-
-
-    $stmt = $mysqli->prepare($sql);
+if(!empty($_GET['sid'])) {
+    $sql = 'SELECT * FROM `members_favourite` WHERE `product_sid`=' . $_GET['sid'];
+    $stmt = $mysqli->query($sql);
+    if ($stmt->num_rows >= 1) {
+        $sql = "DELETE FROM `members_favourite` WHERE `member_sid`=" . $_SESSION['user']['member_sid'] . " AND `product_sid`=" . $_GET['sid'];
 
 
-    $stmt->execute();
-    $af = $stmt->affected_rows;
+        $stmt = $mysqli->prepare($sql);
 
-    $stmt->close();
 
-    if ($af === 1) {
+        $stmt->execute();
+
+
         $result['success'] = true;
+
         unset($_SESSION['love'][$sid]);
+
+
     }
 
+    echo json_encode($result, JSON_UNESCAPED_UNICODE);
 
 }
-
-echo json_encode($result, JSON_UNESCAPED_UNICODE);
-
-
 ?>
